requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cycler==0.11.0
decorator==5.1.0
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
librosa==0.8.1
llvmlite==0.37.0
matplotlib==3.4.3
numba==0.54.1
numpy==1.20.3
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
pooch==1.5.2
pycparser==2.21
pyparsing==2.4.7
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
requests==2.26.0
resampy==0.2.2
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
SoundFile==0.10.3.post1
threadpoolctl==3.0.0
typing-extensions==3.10.0.2
urllib3==1.26.7
